api/requirements.txt
fastapi
uvicorn[standard]
werkzeug
psycopg2-binary
PyJWT
kafka-python